Many fixes

+++ filer/admin/folderadmin.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import unicode_literals
+from __future__ import unicode_literals, absolute_import
 
 import itertools
 import os
@@ -25,26 +25,40 @@ from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext as _
 from django.utils.translation import ungettext, ugettext_lazy
 
-from filer import settings
-from filer.admin.forms import (CopyFilesAndFoldersForm, ResizeImagesForm,
-                               RenameFilesForm)
-from filer.admin.permissions import PrimitivePermissionAwareModelAdmin
-from filer.admin.patched.admin_utils import get_deleted_objects
-from filer.admin.tools import (userperms_for_request,
-                               check_folder_edit_permissions,
-                               check_files_edit_permissions,
-                               check_files_read_permissions,
-                               check_folder_read_permissions,
-                               admin_each_context)
-from filer.models import (Folder, FolderRoot, UnfiledImages, File, tools,
-                          ImagesWithMissingData, FolderPermission, Image)
-from filer.settings import FILER_PAGINATE_BY
-from filer.thumbnail_processors import normalize_subject_location
-from filer.utils.compatibility import (
+from .. import settings
+from .forms import (
+    CopyFilesAndFoldersForm,
+    ResizeImagesForm,
+    RenameFilesForm,
+)
+from .permissions import PrimitivePermissionAwareModelAdmin
+from .patched.admin_utils import get_deleted_objects
+from .tools import (
+    userperms_for_request,
+    check_folder_edit_permissions,
+    check_files_edit_permissions,
+    check_files_read_permissions,
+    check_folder_read_permissions,
+    admin_each_context,
+)
+from .tools import (popup_status, admin_url_params_encoded,
+                    AdminUrlParams)
+from . import views
+from ..models import (
+    FolderRoot,
+    ImagesWithMissingData,
+    UnfiledImages,
+    Folder,
+    File,
+    Image,
+    FolderPermission,
+    tools,
+)
+from ..settings import FILER_PAGINATE_BY
+from ..thumbnail_processors import normalize_subject_location
+from ..utils.compatibility import (
     get_delete_permission, quote, unquote, capfirst)
-from filer.utils.filer_easy_thumbnails import FilerActionThumbnailer
-from filer.views import (popup_status, popup_param, selectfolder_status,
-                         selectfolder_param)
+from ..utils.filer_easy_thumbnails import FilerActionThumbnailer
 
 
 class AddFolderPopupForm(forms.ModelForm):
@@ -121,30 +135,30 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
         Overrides the default to be able to forward to the directory listing
         instead of the default change_list_view
         """
-        r = super(FolderAdmin, self).response_change(request, obj)
-        # Code from django ModelAdmin to determine changelist on the fly
-        if 'Location' in r and r['Location']:
-            # it was a successful save
-            if (r['Location'] in ['../'] or
-                    r['Location'] == self._get_post_url(obj)):
-                if obj.parent:
-                    url = reverse('admin:filer-directory_listing',
-                                  kwargs={'folder_id': obj.parent.id})
-                else:
-                    url = reverse('admin:filer-directory_listing-root')
-                url = "%s%s%s" % (url, popup_param(request),
-                                  selectfolder_param(request, "&"))
-                return HttpResponseRedirect(url)
+        admin_url_params = AdminUrlParams(request)
+        if (
+            request.POST and
+            admin_url_params.popup and
+            admin_url_params.pick and
+            '_continue' not in request.POST
+        ):
+            if obj.parent:
+                url = reverse('admin:filer-directory_listing',
+                              kwargs={'folder_id': obj.parent.id})
             else:
-                # this means it probably was a save_and_continue_editing
-                pass
-        return r
+                url = reverse('admin:filer-directory_listing-root')
+            url = "{}{}".format(
+                url,
+                admin_url_params_encoded(request),
+            )
+            return HttpResponseRedirect(url)
+        return super(FolderAdmin, self).response_change(request, obj)
 
     def render_change_form(self, request, context, add=False, change=False,
                            form_url='', obj=None):
         extra_context = {'show_delete': True,
                          'is_popup': popup_status(request),
-                         'select_folder': selectfolder_status(request), }
+                         'filer_admin_context': AdminUrlParams(request), }
         context.update(extra_context)
         return super(FolderAdmin, self).render_change_form(
             request=request, context=context, add=False,
@@ -159,28 +173,36 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
         before super, because super will delete the object and make it
         impossible to find out the parent folder to redirect to.
         """
-        parent_folder = None
         try:
             obj = self.get_queryset(request).get(pk=unquote(object_id))
             parent_folder = obj.parent
         except self.model.DoesNotExist:
-            obj = None
-
-        r = super(FolderAdmin, self).delete_view(
-            request=request, object_id=object_id,
-            extra_context=extra_context)
-
-        url = r.get("Location", None)
-        if url in ["../../../../", "../../"] or url == self._get_post_url(obj):
+            parent_folder = None
+        admin_url_params = AdminUrlParams(request)
+        if (
+            request.POST and
+            admin_url_params.popup and
+            admin_url_params.pick
+        ):
+            # Popup in pick mode. Call super delete view so the objects
+            # actually get deleted. All possible failures in delete_view cause
+            # exceptions, so it is safe to ignore the return value though.
+            super(FolderAdmin, self).delete_view(
+                request=request, object_id=object_id,
+                extra_context=extra_context)
             if parent_folder:
                 url = reverse('admin:filer-directory_listing',
                               kwargs={'folder_id': parent_folder.id})
             else:
                 url = reverse('admin:filer-directory_listing-root')
-            url = "%s%s%s" % (url, popup_param(request),
-                              selectfolder_param(request, "&"))
+            url = "{}{}".format(
+                url,
+                admin_url_params_encoded(request),
+            )
             return HttpResponseRedirect(url)
-        return r
+        return super(FolderAdmin, self).delete_view(
+            request=request, object_id=object_id,
+            extra_context=extra_context)
 
     def icon_img(self, xs):
         return mark_safe(('<img src="%simg/icons/plainfolder_32x32.png" '
@@ -190,7 +212,6 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
     def get_urls(self):
         from django.conf.urls import patterns, url
         urls = super(FolderAdmin, self).get_urls()
-        from filer import views
         url_patterns = patterns('',
             # we override the default list view with our own directory listing
             # of the root directories
@@ -240,10 +261,10 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
                 Folder.objects.get(id=last_folder_id)
             except Folder.DoesNotExist:
                 url = reverse('admin:filer-directory_listing-root')
-                url = "%s%s%s" % (url, popup_param(request), selectfolder_param(request, "&"))
+                url = "%s%s" % (url, admin_url_params_encoded(request))
             else:
                 url = reverse('admin:filer-directory_listing', kwargs={'folder_id': last_folder_id})
-                url = "%s%s%s" % (url, popup_param(request), selectfolder_param(request, "&"))
+                url = "%s%s" % (url, admin_url_params_encoded(request))
             return HttpResponseRedirect(url)
         elif folder_id is None:
             folder = FolderRoot()
@@ -409,7 +430,7 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
             'folder_files': folder_files,
             'limit_search_to_folder': limit_search_to_folder,
             'is_popup': popup_status(request),
-            'select_folder': selectfolder_status(request),
+            'filer_admin_context': AdminUrlParams(request),
             # needed in the admin/base.html template for logout links
             'root_path': reverse('admin:index'),
             'action_form': action_form,
@@ -747,7 +768,7 @@ class FolderAdmin(PrimitivePermissionAwareModelAdmin):
             "protected": all_protected,
             "opts": opts,
             'is_popup': popup_status(request),
-            'select_folder': selectfolder_status(request),
+            'filer_admin_context': AdminUrlParams(request),
             "root_path": reverse('admin:index'),
             "app_label": app_label,
             "action_checkbox_name": helpers.ACTION_CHECKBOX_NAME,
